# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Fundonotes API",
      "version": "1.0.0",
      "description": "API Documentation for Fundonotes App"
    },
    "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
    "paths": {
      "/api/v1/users/register": {
        "post": {
          "summary": "Create a new user",
          "operationId": "registerUser",
          "tags": ["Users"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string" }
                  },
                  "required": ["firstName", "lastName", "email", "password"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User registered successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": { "description": "Bad request" }
          }
        }
      },
      "/api/v1/users/login": {
        "post": {
          "summary": "Login a user",
          "operationId": "loginUser",
          "tags": ["Users"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": { "type": "string" },
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/v1/users/forgetPassword": {
        "post": {
          "summary": "Request a password reset",
          "operationId": "forgetPassword",
          "tags": ["Users"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "format": "email" }
                  },
                  "required": ["email"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Password reset request sent successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": { "description": "Bad request" }
          }
        }
      },
      "/api/v1/users/resetPassword": {
        "put": {
          "summary": "Reset user password",
          "operationId": "resetPassword",
          "tags": ["Users"],
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Password reset successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": { "description": "Bad request" }
          }
        }
      },
      "/api/v1/users/refreshToken": {
        "get": {
          "summary": "Refresh user access token",
          "operationId": "refreshToken",
          "tags": ["Users"],
          "security": [{ "bearerAuth": [] }],
          "responses": {
            "200": {
              "description": "Token refreshed successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": { "type": "string" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/v1/notes": {
        "post": {
          "summary": "Create a new note",
          "operationId": "createNote",
          "tags": ["Notes"],
          "security": [{ "bearerAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "description": { "type": "string" },
                    "color": { "type": "string" }
                  },
                  "required": ["title"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Note created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": { "description": "Bad request" }
          }
        },
        "get": {
            "summary": "Get all notes with pagination",
            "operationId": "getAllNotes",
            "tags": ["Notes"],
            "security": [{ "bearerAuth": [] }],
            "parameters": [
              {
                "name": "page",
                "in": "query",
                "description": "Page number (must be a positive integer)",
                "required": false,
                "schema": {
                  "type": "integer",
                  "default": 1,
                  "minimum": 1
                }
              },
              {
                "name": "limit",
                "in": "query",
                "description": "Number of notes per page (must be a positive integer)",
                "required": false,
                "schema": {
                  "type": "integer",
                  "default": 5,
                  "minimum": 1
                }
              }
            ],
            "responses": {
              "200": {
                "description": "List of notes fetched successfully",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "integer", "example": 200 },
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "title": { "type": "string", "example": "Sample Note Title" },
                              "description": { "type": "string", "example": "This is a sample note description." },
                              "color": { "type": "string", "example": "#FFFFFF" }
                            }
                          }
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "page": { "type": "integer", "example": 1 },
                            "limit": { "type": "integer", "example": 5 },
                            "totalRecords": { "type": "integer", "example": 50 },
                            "totalPages": { "type": "integer", "example": 10 }
                          }
                        },
                        "message": { "type": "string", "example": "Successfully fetched all notes" }
                      }
                    }
                  }
                }
              },
              "400": {
                "description": "Bad request (e.g., invalid page or limit values)",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "integer", "example": 400 },
                        "message": { "type": "string", "example": "Page and limit must be positive integers" }
                      }
                    }
                  }
                }
              },
              "401": {
                "description": "Unauthorized (invalid or missing Bearer token)",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "code": { "type": "integer", "example": 401 },
                        "message": { "type": "string", "example": "Unauthorized access" }
                      }
                    }
                  }
                }
              }
            }
          }
       },
      "/api/v1/notes/{noteId}": {
        "get": {
          "summary": "Get a specific note",
          "operationId": "getNote",
          "tags": ["Notes"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "name": "noteId",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "Note found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "color": { "type": "string" }
                    }
                  }
                }
              }
            },
            "404": { "description": "Note not found" }
          }
        },
  
  "put": {
    "summary": "Update a specific note",
    "operationId": "updateNote",
    "tags": ["Notes"],
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "noteId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "title": { "type": "string" },
              "description": { "type": "string" },
              "color": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Note updated successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "400": { "description": "Bad request" }
  }
},

        "delete": {
          "summary": "Delete a specific note",
          "operationId": "permanentlyDeleteNote",
          "tags": ["Notes"],
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            {
              "name": "noteId",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "Note deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "404": { "description": "Note not found" }
          }
        }
     },
  "/api/v1/notes/{noteId}/trash": {
  "put": {
    "summary": "Trash a specific note",
    "operationId": "trashNote",
    "tags": ["Notes"],
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "noteId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "responses": {
      "200": {
        "description": "Note trashed successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "404": { "description": "Note not found" }
    }
  }
},
  "/api/v1/notes/{noteId}/archive": {
  "put": {
    "summary": "Archive a specific note",
    "operationId": "archiveNote",
    "tags": ["Notes"],
    "security": [{ "bearerAuth": [] }],
    "parameters": [
      {
        "name": "noteId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "responses": {
      "200": {
        "description": "Note archived successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "404": { "description": "Note not found" }
    }
  }
}
}
} 